streamlit/requirements.txt
streamlit
snowflake-connector-python
cryptography
pandas
